pyproject.toml
[tool.poetry]
name = "biblioteka"
version = "0.1.0"
description = "aplicação que faz a gestão de uma biblioteca"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.2.2"
django = "^4.2"
ipdb = "^0.13.13"
drf-spectacular = "^0.26.2"
django-apscheduler = "^0.6.2"
dj-database-url = "^2.0.0"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
